requirements.txt
flask>=3.1.1
python-telegram-bot[webhooks]==20.7
httpx>=0.25.2
anyio>=4.9.0
certifi>=2025.4.26